Moved common ui code to core-ui module
Moved image picker Bottomsheet code to core-ui module
Moved draw utils and height spacer to core-ui module
Added shared compose groovy gradle plugin
