+++ core/core-ui/src/main/java/com/rempawl/bottomsheet/GalleryPickerOption.kt
@@ -1,14 +1,13 @@
-package com.rempawl.image.processing.core
+package com.rempawl.bottomsheet
 
 import androidx.activity.result.PickVisualMediaRequest
 import androidx.activity.result.contract.ActivityResultContracts
-import com.rempawl.image.processing.core.GalleryPickerOption.IMAGE_ONLY
+import com.rempawl.bottomsheet.GalleryPickerOption.IMAGE_ONLY
 
 enum class GalleryPickerOption {
     IMAGE_ONLY;
 }
 
-// todo ui-core module
 fun GalleryPickerOption.toPickVisualMediaRequest(): PickVisualMediaRequest =
     when (this) {
         IMAGE_ONLY -> PickVisualMediaRequest(ActivityResultContracts.PickVisualMedia.ImageOnly)
