+++ core/core-ui/src/main/res/drawable/ic_camera.xml
