+++ core/core-ui/src/main/res/drawable/ic_gallery.xml
