+++ core/core-ui/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="image_source_picker_title">Select image source</string>
+    <string name="image_source_picker_camera_label">Camera</string>
+    <string name="image_source_picker_gallery_label">Gallery</string>
+</resources>
