+++ feature/image-processing/src/main/java/com/rempawl/image/processing/viewmodel/ImageProcessingAction.kt
@@ -1,7 +1,7 @@
 package com.rempawl.image.processing.viewmodel
 
 import com.rempawl.core.viewmodel.mvi.Action
-import com.rempawl.image.processing.core.ImageSourcePickerOption
+import com.rempawl.bottomsheet.ImageSourcePickerOption
 
 sealed interface ImageProcessingAction : Action {
     data object SelectImageFabClicked : ImageProcessingAction
