+++ feature/image-processing/src/main/java/com/rempawl/image/processing/viewmodel/ImageProcessingEffect.kt
@@ -1,7 +1,7 @@
 package com.rempawl.image.processing.viewmodel
 
 import com.rempawl.core.viewmodel.mvi.Effect
-import com.rempawl.image.processing.core.GalleryPickerOption
+import com.rempawl.bottomsheet.GalleryPickerOption
 
 sealed interface ImageProcessingEffect : Effect {
     data class TakePicture(val uri: String) : ImageProcessingEffect
