+++ feature/image-processing/src/main/res/values/strings.xml
@@ -4,7 +4,4 @@
     <string name="title_objects_section">Detected objects</string>
     <string name="title_texts_section">Detected texts</string>
     <string name="error_generic">An error ocurred</string>
-    <string name="image_source_picker_title">Select image source</string>
-    <string name="image_source_picker_camera_label">Camera</string>
-    <string name="image_source_picker_gallery_label">Gallery</string>
 </resources>
