+++ plugin-shared/src/main/groovy/com.rempawl.shared.feature.gradle
@@ -1,23 +1,9 @@
 plugins {
-    id "com.rempawl.shared.android"
-    id "com.google.devtools.ksp"
-    id "com.rempawl.ktlint"
-    id "kotlin-parcelize"
-    id("org.jetbrains.kotlin.plugin.compose")
-}
-
-android {
-    buildFeatures {
-        compose = true
-    }
-    composeOptions {
-        kotlinCompilerExtensionVersion = libs.versions.composeCompiler.get()
-    }
-}
-
-composeCompiler {
-    reportsDestination = layout.buildDirectory.dir("compose_compiler")
-//    stabilityConfigurationFile = rootProject.layout.projectDirectory.file("stability_config.conf") todo
+    id("com.rempawl.shared.android")
+    id("com.rempawl.shared.compose")
+    id("com.google.devtools.ksp")
+    id("com.rempawl.ktlint")
+    id("kotlin-parcelize")
 }
 
 dependencies {
@@ -26,11 +12,10 @@ dependencies {
     implementation(project(":core:core-ui"))
     implementation(project(":core:core-viewmodel"))
 
-    implementation(platform(libs.androidx.compose.bom))
     implementation(libs.androidx.activity.compose)
 
     implementation(libs.androidx.runtime.tracing)
-    implementation(libs.androidx.material3.android)
+
     implementation(libs.androidx.lifecycle.runtime.compose.android)
     implementation(libs.androidx.lifecycle.viewmodel.android)
